+++ src/lib.rs
@@ -19,4 +19,5 @@ pub mod prefs;
 pub mod result_heuristic;
 pub mod run_log;
 pub mod status;
+pub mod type_override;
 pub mod unicode_relocation;